requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
django-cors-headers>=4.4.0,<5.0
psycopg2-binary>=2.9.0,<3.0
python-dotenv>=1.0.0,<2.0
